First commit on mock scheduler
The Python implementation is functionnal.  It takes HTTP requests made
by mock_ord_soumet.

The rust implementation is just the beginning of an HTTP echo server.
I spent a good deal of time trying to find an HTTP library but there are
like a million choices and no "standard" one.

So I looked at what could be done with the standard library and it looks
like the only thing that is part of the standard library is TCP stuff
and after following the Rust book, I quite the time figuring out how to
get the body of the request.